% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kobo_flatten}
\alias{kobo_flatten}
\title{Recursively flatten a nested dataframe from Kobo.}
\usage{
kobo_flatten(x, ignore = "^_")
}
\arguments{
\item{x}{Dataframe to flatten}

\item{ignore}{Regular expression to match columsn to ignore}
}
\value{
A flattened dataframe
}
\description{
This function is for internal use in the package only.
It takes a nested dataframe and recursively flattens the
nested tables. It will ignore columns with names defined by the 'ignore'
parameter.
}
