% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{kobor_assets}
\alias{kobor_assets}
\title{Read list of assets from Kobo toolbox}
\usage{
kobor_assets(username, password)
}
\arguments{
\item{username}{Your Kobo username}

\item{password}{Your Kobo password}
}
\value{
A dataframe of asset (form) details.
}
\description{
Connects to Kobo toolbox and downloads details of all
assets (forms). The  uid's are useful as they provide
the idnetifier for downloading individual asset/form data.
Note: We strongly advise against storing your Kobo password in a script.
}
