% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{kobor_read}
\alias{kobor_read}
\title{Download an asset's (form) data from Kobo toolbox.}
\usage{
kobor_read(asset, username, password)
}
\arguments{
\item{asset}{The id for the data you wish to download}

\item{username}{Your Kobo username}

\item{password}{Your Kobo password}
}
\value{
The downloaded data in flattened format.
}
\description{
Connects to Kobo toolbox and downloads data.
Needs the id of the asset (form). You can obtain this using kobo_assets()
It's the uid that you use to identify the form.
Note: We strongly advise against storing your Kobo password in a script.
}
